[kafka_consumer][test] more pytest changes
