+++ kafka_consumer/tox.ini
@@ -2,7 +2,7 @@
 [tox]
 minversion = 2.0
 basepython = py27
-envlist = unit, kafkaconsumer{_0811,_0901,_0901_zk,_01021,_01021_zk,_01101,_01101_zk}, flake8
+envlist = kafkaconsumer{_0811,_0901,_0901_zk,_01021,_01021_zk,_01101,_01101_zk}, flake8
 
 [common]
 deps =
@@ -14,25 +14,28 @@ commands =
 
 [testenv]
 platform = linux2|darwin
+
+[testenv:common]
 setenv = 
     EXTERNAL_JMX_PORT=9999
     EXTERNAL_PORT=9092
     ZOOKEEPER_VERSION=3.4.9
-    KAFKA_CREATE_TOPICS="marvel:2:1,dc:2:1"
+    KAFKA_CREATE_TOPICS=marvel:2:1,dc:2:1
     JMX_PORT=9999
-    KAFKA_HEAP_OPTS="-Xmx256M -Xms128M"
+    KAFKA_HEAP_OPTS=-Xmx256M -Xms128M
+    KAFKA_ZOOKEEPER_CONNECT=zookeeper:2181
     KAFKA_ADVERTISED_HOST_NAME="172.17.0.1"
-    KAFKA_ZOOKEEPER_CONNECT="zookeeper:2181"
 
-[testenv:unit]
-deps = {[common]deps}
-commands =
-    pip install --require-hashes -r requirements.txt
-    pytest -m"not integration" -v
+# [testenv:unit]
+# deps = {[common]deps}
+# commands =
+#     pip install --require-hashes -r requirements.txt
+#     pytest -m"not integration" -v
 
 [testenv:kafkaconsumer_0811]
 deps = {[common]deps}
 setenv = 
+    {[testenv:common]setenv}
     KAFKA_VERSION=0.8.1.1-1
     KAFKA_OFFSETS_STORAGE=zookeeper
 passenv = DOCKER_HOST*
@@ -41,6 +44,7 @@ commands = {[common]commands}
 [testenv:kafkaconsumer_0901]
 deps = {[common]deps}
 setenv = 
+    {[testenv:common]setenv}
     KAFKA_VERSION=0.9.0.1-1
     KAFKA_OFFSETS_STORAGE=kafka
 passenv = DOCKER_HOST*
@@ -49,6 +53,7 @@ commands = {[common]commands}
 [testenv:kafkaconsumer_0901_zk]
 deps = {[common]deps}
 setenv = 
+    {[testenv:common]setenv}
     KAFKA_VERSION=0.9.0.1-1
     KAFKA_OFFSETS_STORAGE=zookeeper
 passenv = DOCKER_HOST*
@@ -57,6 +62,7 @@ commands = {[common]commands}
 [testenv:kafkaconsumer_01021]
 deps = {[common]deps}
 setenv =
+    {[testenv:common]setenv}
     KAFKA_VERSION=0.10.2.1
     KAFKA_OFFSETS_STORAGE=kafka
 passenv = DOCKER_HOST*
@@ -65,6 +71,7 @@ commands = {[common]commands}
 [testenv:kafkaconsumer_01021_zk]
 deps = {[common]deps}
 setenv = 
+    {[testenv:common]setenv}
     KAFKA_VERSION=0.10.2.1
     KAFKA_OFFSETS_STORAGE=zookeeper
 passenv = DOCKER_HOST*
@@ -73,6 +80,7 @@ commands = {[common]commands}
 [testenv:kafkaconsumer_01101]
 deps = {[common]deps}
 setenv = 
+    {[testenv:common]setenv}
     KAFKA_VERSION=0.11.0.1
     KAFKA_OFFSETS_STORAGE=kafka
 passenv = DOCKER_HOST*
@@ -81,6 +89,7 @@ commands = {[common]commands}
 [testenv:kafkaconsumer_01101_zk]
 deps = {[common]deps}
 setenv = 
+    {[testenv:common]setenv}
     KAFKA_VERSION=0.11.0.1
     KAFKA_OFFSETS_STORAGE=zookeeper
 passenv = DOCKER_HOST*
